Array DSL slice 3: sync-condition DI leaf + array-op filter
Close matrix row: ReactiveArray.Where(pred) -> array-op{op:filter, predicate} ->
runtime filters by per-element sync predicate.

- Import-cycle resolution (§14.1): extract sync condition subset (compare/all/any/
  not) into conditions/sync-condition.ts as a leaf that receives the value-evaluator
  by DI (evalValue param) and imports nothing from core/evaluate. conditions.ts
  delegates its sync subset + async lane to it (keeps public API). evaluate.ts
  imports the leaf for per-element predicates. No cycle; one resolver preserved.
- C#: ArrayOperationExpression.Predicate (ValidationCondition sync subset, nullable
  by design per Rule 6 — absent for count, present for filter); ArrayFilter factory.
  Generator gains Optional() field support; ArrayOp += filter; predicate? declared.
- Runtime: filter case pushes withElement(item) per element, evaluates the sync
  predicate via the leaf, collects matches; immediate lane (no Promise).
- 3 new vitest (object-member predicate, x=>x primitive self-read, DOMTokenList).
  156 total green incl. all 14 conditions tests; typecheck clean.
