+++ Alis.Reactive.Assets/runtime/conditions/conditions.ts
@@ -1,53 +1,22 @@
 // conditions.ts — V3 ConditionGraph evaluation.
-// Uses SHARED resolver for value resolution.
-// ConditionGraph is a discriminated union: compare, all, any, not, confirm.
-// ValidationCondition is the sync subset: compare, all, any, not.
+// The SYNC subset (compare/all/any/not) lives in ./sync-condition, a DI leaf that
+// receives the value-evaluator as a parameter (so it never imports core/evaluate). This
+// module keeps the public entry points and owns the async lane: confirm is the only term
+// that crosses to async, so full ConditionGraph evaluation may return a Promise.
 
 import type {
   ConditionGraph,
-  CompareCondition,
-  CollectionItemCompareCondition,
-  EqualityCompareOp,
-  EqualityCompareCondition,
-  MembershipCompareCondition,
-  MembershipCompareOp,
-  OrderedCompareOp,
-  OrderedCompareCondition,
   PlanDocument,
-  RangeCompareCondition,
-  RegexCompareCondition,
-  Shape,
-  TextLengthCompareCondition,
-  TextCompareOp,
-  TextCompareCondition,
-  UnaryCompareOp,
-  ValueExpression,
+  ExecContext,
   ValidationCondition,
 } from "../types";
-import type { ExecContext } from "../types";
 import { scope } from "../core/trace";
 import { assertNever } from "../core/assert-never";
-import { applyShape, toString } from "../core/shape-convert";
 import { evaluateValue } from "../core/evaluate";
 import { ExecutionContext } from "../domain/execution-context";
-import { RuntimeShape } from "../domain/runtime-shape";
+import { evaluateSyncCondition, evaluateCompare } from "./sync-condition";
 
 const log = scope("conditions");
-
-type TextOperand =
-  | { readonly kind: "text"; readonly value: string }
-  | { readonly kind: "missing" };
-
-interface ComparisonLeft {
-  readonly raw: unknown;
-  readonly shaped: unknown;
-}
-
-type OrderedConditionValue =
-  | { readonly kind: "number"; readonly value: number }
-  | { readonly kind: "text"; readonly value: string }
-  | { readonly kind: "boolean"; readonly value: boolean }
-  | { readonly kind: "missing" };
 
 interface AlisBrowserApi {
   readonly alis?: {
@@ -55,13 +24,9 @@ interface AlisBrowserApi {
   };
 }
 
-const missingText: TextOperand = { kind: "missing" };
-const missingOrderedConditionValue: OrderedConditionValue = { kind: "missing" };
-const noRightOperandTrace = { kind: "none" } as const;
-
-/** Sync condition evaluation for validation conditions. */
+/** Sync condition evaluation for validation conditions (compare/all/any/not). */
 export function evaluateCondition(condition: ValidationCondition, plan: PlanDocument, ctx?: ExecContext): boolean {
-  return evaluateConditionSync(condition, plan, ExecutionContext.from(ctx));
+  return evaluateSyncCondition(condition, plan, ExecutionContext.from(ctx), evaluateValue);
 }
 
 /** Current-lane condition evaluation. Crosses to async only when a reached term requires it. */
@@ -73,25 +38,6 @@ export function evaluateConditionInCurrentLane(
   return evaluateConditionInLane(condition, plan, ExecutionContext.from(ctx));
 }
 
-function evaluateConditionSync(
-  condition: ValidationCondition,
-  plan: PlanDocument,
-  context: ExecutionContext,
-): boolean {
-  switch (condition.kind) {
-    case "compare":
-      return evaluateCompare(condition, plan, context);
-    case "all":
-      return condition.terms.every(term => evaluateConditionSync(term, plan, context));
-    case "any":
-      return condition.terms.some(term => evaluateConditionSync(term, plan, context));
-    case "not":
-      return !evaluateConditionSync(condition.term, plan, context);
-    default:
-      return assertNever(condition, "condition kind");
-  }
-}
-
 function evaluateConditionInLane(
   condition: ConditionGraph,
   plan: PlanDocument,
@@ -99,7 +45,7 @@ function evaluateConditionInLane(
 ): boolean | Promise<boolean> {
   switch (condition.kind) {
     case "compare":
-      return evaluateCompare(condition, plan, context);
+      return evaluateCompare(condition, plan, context, evaluateValue);
     case "all":
       return evaluateAllInLane(condition.terms, plan, context, 0);
     case "any":
@@ -172,303 +118,4 @@ async function evaluateConfirmCondition(message: string): Promise<boolean> {
   const accepted = await confirmFn(message);
   log.debug("confirm.result", { accepted, message });
   return accepted;
-}
-
-// -- Compare evaluation --
-
-function evaluateCompare(condition: CompareCondition, plan: PlanDocument, context: ExecutionContext): boolean {
-  const left = resolveComparisonLeft(condition, plan, context);
-
-  switch (condition.op) {
-    case "is-null":
-    case "not-null":
-    case "is-empty":
-    case "not-empty":
-    case "truthy":
-    case "falsy":
-      traceCompare(condition, left, noRightOperandTrace);
-      return unaryMatches(condition.op, left);
-
-    case "eq":
-    case "neq": {
-      const right = resolveRightValue(condition, plan, context, condition.shape);
-      traceCompare(condition, left, right);
-      return equalityMatches(condition.op, left, right);
-    }
-
-    case "gt":
-    case "gte":
-    case "lt":
-    case "lte": {
-      const right = resolveRightValue(condition, plan, context, condition.shape);
-      traceCompare(condition, left, right);
-      return orderedComparisonMatches(condition.op, left.shaped, right);
-    }
-
-    case "in":
-    case "not-in": {
-      const right = resolveMembershipItems(condition, plan, context);
-      traceCompare(condition, left, right);
-      return membershipMatches(condition.op, left, right);
-    }
-
-    case "between": {
-      const [lower, upper] = resolveRangeBounds(condition, plan, context);
-      traceCompare(condition, left, [lower, upper]);
-      return inclusiveRangeContains(lower, upper, left.shaped);
-    }
-
-    case "array-contains": {
-      const right = resolveRightValue(condition, plan, context, condition.itemShape);
-      const items = shapeCollectionItems(left.shaped, condition.itemShape);
-      traceCompare(condition, left, right);
-      return items !== undefined && items.includes(right);
-    }
-
-    case "contains":
-      return evaluateTextComparison(condition.op, left, textOperand(condition), condition);
-
-    case "starts-with":
-      return evaluateTextComparison(condition.op, left, textOperand(condition), condition);
-
-    case "ends-with":
-      return evaluateTextComparison(condition.op, left, textOperand(condition), condition);
-
-    case "matches":
-      return evaluateRegexComparison(left, textOperand(condition), condition);
-
-    case "min-length":
-      return evaluateMinimumLengthComparison(left, minimumTextLength(condition), condition);
-
-    default:
-      return assertNever(condition, "compare condition");
-  }
-}
-
-function resolveComparisonLeft(condition: CompareCondition, plan: PlanDocument, context: ExecutionContext): ComparisonLeft {
-  const raw = evaluateValue(condition.left, plan, context.raw);
-  const shaped = applyShape(raw, condition.shape);
-  return { raw, shaped };
-}
-
-type ScalarRightCondition =
-  | EqualityCompareCondition
-  | OrderedCompareCondition
-  | CollectionItemCompareCondition;
-
-type TextRightCondition =
-  | TextCompareCondition
-  | RegexCompareCondition;
-
-function traceCompare(condition: CompareCondition, left: ComparisonLeft, right: unknown): void {
-  log.trace("compare", { op: condition.op, left: left.shaped, right });
-}
-
-function resolveRightValue(
-  condition: ScalarRightCondition,
-  plan: PlanDocument,
-  context: ExecutionContext,
-  shape: Shape,
-): unknown {
-  const raw = evaluateValue(condition.right.value, plan, context.raw);
-  return applyShape(raw, shape);
-}
-
-function resolveMembershipItems(
-  condition: MembershipCompareCondition,
-  plan: PlanDocument,
-  context: ExecutionContext,
-): unknown[] {
-  return condition.right.value.items.map(item =>
-    resolveShapedComparisonItem(item, plan, context, condition.shape));
-}
-
-function resolveRangeBounds(
-  condition: RangeCompareCondition,
-  plan: PlanDocument,
-  context: ExecutionContext,
-): [unknown, unknown] {
-  const [lower, upper] = condition.right.value.items;
-  return [
-    resolveShapedComparisonItem(lower, plan, context, condition.shape),
-    resolveShapedComparisonItem(upper, plan, context, condition.shape),
-  ];
-}
-
-function resolveShapedComparisonItem(
-  producer: ValueExpression,
-  plan: PlanDocument,
-  context: ExecutionContext,
-  shape: Shape,
-): unknown {
-  return applyShape(evaluateValue(producer, plan, context.raw), shape);
-}
-
-function textOperand(condition: TextRightCondition): string {
-  return condition.right.value.value;
-}
-
-function minimumTextLength(condition: TextLengthCompareCondition): number {
-  return condition.right.value.value;
-}
-
-function unaryMatches(op: UnaryCompareOp, left: ComparisonLeft): boolean {
-  switch (op) {
-    case "is-null":
-      return isMissingValue(left.raw);
-    case "not-null":
-      return !isMissingValue(left.raw);
-    case "is-empty":
-      return isEmpty(left.raw);
-    case "not-empty":
-      return !isEmpty(left.raw);
-    case "truthy":
-      return !!left.shaped;
-    case "falsy":
-      return !left.shaped;
-    default:
-      return assertNever(op, "unary comparison operator");
-  }
-}
-
-function equalityMatches(op: EqualityCompareOp, left: ComparisonLeft, right: unknown): boolean {
-  const valuesAreEqual = left.shaped === right;
-  if (op === "eq") return valuesAreEqual;
-
-  return !valuesAreEqual;
-}
-
-function membershipMatches(op: MembershipCompareOp, left: ComparisonLeft, values: readonly unknown[]): boolean {
-  const collectionContainsLeft = values.includes(left.shaped);
-  if (op === "in") return collectionContainsLeft;
-
-  return !collectionContainsLeft;
-}
-
-function shapeCollectionItems(value: unknown, itemShape: Shape): unknown[] | undefined {
-  if (!Array.isArray(value)) return undefined;
-
-  return RuntimeShape.from(itemShape).applyEach(value);
-}
-
-function evaluateTextComparison(
-  op: TextCompareOp,
-  left: ComparisonLeft,
-  right: string,
-  condition: CompareCondition,
-): boolean {
-  traceCompare(condition, left, right);
-
-  switch (op) {
-    case "contains":
-      return textOp(left.shaped, right, (source, operand) => source.includes(operand));
-    case "starts-with":
-      return textOp(left.shaped, right, (source, operand) => source.startsWith(operand));
-    case "ends-with":
-      return textOp(left.shaped, right, (source, operand) => source.endsWith(operand));
-    default:
-      return assertNever(op, "text comparison operator");
-  }
-}
-
-function evaluateRegexComparison(left: ComparisonLeft, pattern: string, condition: CompareCondition): boolean {
-  traceCompare(condition, left, pattern);
-
-  const leftText = asText(left.shaped);
-  if (leftText.kind === "missing") return false;
-
-  return new RegExp(pattern).test(leftText.value);
-}
-
-function evaluateMinimumLengthComparison(left: ComparisonLeft, minimumLength: number, condition: CompareCondition): boolean {
-  traceCompare(condition, left, minimumLength);
-
-  const leftText = asText(left.shaped);
-  return leftText.kind === "text" && leftText.value.length >= minimumLength;
-}
-
-function orderedComparisonMatches(op: OrderedCompareOp, left: unknown, right: unknown): boolean {
-  const comparison = compareOrderedValues(left, right);
-  if (comparison === undefined) return false;
-
-  switch (op) {
-    case "gt":
-      return comparison > 0;
-    case "gte":
-      return comparison >= 0;
-    case "lt":
-      return comparison < 0;
-    case "lte":
-      return comparison <= 0;
-    default:
-      assertNever(op, "ordered comparison operator");
-  }
-}
-
-function inclusiveRangeContains(lower: unknown, upper: unknown, value: unknown): boolean {
-  const boundsCanBeOrdered = compareOrderedValues(lower, upper) !== undefined;
-  if (!boundsCanBeOrdered) return false;
-
-  const lowerComparison = compareOrderedValues(value, lower);
-  const upperComparison = compareOrderedValues(value, upper);
-  if (lowerComparison === undefined || upperComparison === undefined) return false;
-
-  return lowerComparison >= 0 && upperComparison <= 0;
-}
-
-function compareOrderedValues(left: unknown, right: unknown): number | undefined {
-  const leftValue = toOrderedConditionValue(left);
-  const rightValue = toOrderedConditionValue(right);
-
-  switch (leftValue.kind) {
-    case "number":
-      return rightValue.kind === "number" ? leftValue.value - rightValue.value : undefined;
-    case "text":
-      if (rightValue.kind !== "text") return undefined;
-      if (leftValue.value === rightValue.value) return 0;
-      return leftValue.value > rightValue.value ? 1 : -1;
-    case "boolean":
-      return rightValue.kind === "boolean"
-        ? booleanRank(leftValue.value) - booleanRank(rightValue.value)
-        : undefined;
-    case "missing":
-      return undefined;
-    default:
-      return assertNever(leftValue, "ordered condition value");
-  }
-}
-
-function toOrderedConditionValue(value: unknown): OrderedConditionValue {
-  if (typeof value === "number" && Number.isFinite(value)) return { kind: "number", value };
-  if (typeof value === "string") return { kind: "text", value };
-  if (typeof value === "boolean") return { kind: "boolean", value };
-
-  return missingOrderedConditionValue;
-}
-
-function booleanRank(value: boolean): number {
-  return value ? 1 : 0;
-}
-
-function textOp(left: unknown, right: string, predicate: (source: string, operand: string) => boolean): boolean {
-  const leftText = asText(left);
-  return leftText.kind === "text" && predicate(leftText.value, right);
-}
-
-function isEmpty(value: unknown): boolean {
-  const valueIsEmptyText = value === "";
-  const valueIsMissing = isMissingValue(value);
-  const valueIsEmptyCollection = Array.isArray(value) && value.length === 0;
-  return valueIsEmptyText || valueIsMissing || valueIsEmptyCollection;
-}
-
-function asText(value: unknown): TextOperand {
-  if (isMissingValue(value)) return missingText;
-  const result = toString(value);
-  if (result.ok) return { kind: "text", value: result.value };
-  return missingText;
-}
-
-function isMissingValue(value: unknown): boolean {
-  return value === null || value === undefined;
 }
